pyproject.toml
[project]
name = "Back-End"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AndreRicardoSc", email = "andrenarutto7@gmail.com"},
]
dependencies = ["django>=4.2.23", "djangorestframework>=3.16.0", "Pillow>=11.3.0", "python-magic>=0.4.27", "djangorestframework-simplejwt>=5.5.0", "django-cors-headers>=4.8.0", "dj-database-url==3.0.1", "whitenoise==6.10.0", "gunicorn>=23.0.0", "waitress>=3.0.2", "psycopg2-binary>=2.9.10", "django-storages>=1.14.6", "boto3>=1.40.33"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = true


[tool.pdm]
distribution = true
